+++ sdk/go/kubernetes/provider.go
@@ -25,9 +25,6 @@ func NewProvider(ctx *pulumi.Context,
 	if args.EnableConfigMapMutable == nil {
 		args.EnableConfigMapMutable = pulumi.BoolPtr(getEnvOrDefault(false, parseEnvBool, "PULUMI_K8S_ENABLE_CONFIGMAP_MUTABLE").(bool))
 	}
-	if args.EnableDryRun == nil {
-		args.EnableDryRun = pulumi.BoolPtr(getEnvOrDefault(false, parseEnvBool, "PULUMI_K8S_ENABLE_DRY_RUN").(bool))
-	}
 	if args.EnableReplaceCRD == nil {
 		args.EnableReplaceCRD = pulumi.BoolPtr(getEnvOrDefault(false, parseEnvBool, "PULUMI_K8S_ENABLE_REPLACE_CRD").(bool))
 	}
@@ -71,10 +68,6 @@ type providerArgs struct {
 	// 1. This `enableConfigMapMutable` parameter.
 	// 2. The `PULUMI_K8S_ENABLE_CONFIGMAP_MUTABLE` environment variable.
 	EnableConfigMapMutable *bool `pulumi:"enableConfigMapMutable"`
-	// Deprecated. If present and set to true, enable server-side diff calculations.
-	//
-	// Deprecated: This option has been replaced by `enableServerSideApply`.
-	EnableDryRun *bool `pulumi:"enableDryRun"`
 	// Obsolete. This option has no effect.
 	//
 	// Deprecated: This option is deprecated, and will be removed in a future release.
@@ -126,10 +119,6 @@ type ProviderArgs struct {
 	// 1. This `enableConfigMapMutable` parameter.
 	// 2. The `PULUMI_K8S_ENABLE_CONFIGMAP_MUTABLE` environment variable.
 	EnableConfigMapMutable pulumi.BoolPtrInput
-	// Deprecated. If present and set to true, enable server-side diff calculations.
-	//
-	// Deprecated: This option has been replaced by `enableServerSideApply`.
-	EnableDryRun pulumi.BoolPtrInput
 	// Obsolete. This option has no effect.
 	//
 	// Deprecated: This option is deprecated, and will be removed in a future release.